+++ src/views/datatable-cruds.blade.php
@@ -1,4 +1,17 @@
-@extends($extends, $extendsData)
-@section($section)
+@extends($extendsName, $extendsData)
+
+@foreach ($sections as $section)
+    @section($section['name'])
+    {{ $section['value'] }}
+    @endsection
+@endforeach
+
+@foreach ($stacks as $stack)
+    @push($stack['name'])
+    {{ $stack['value'] }}
+    @endpush
+@endforeach
+
+@section($sectionName)
 @datatable($datatable)
 @endsection